Moved generic code from the lwjgl to core package
